+++ ui/src/pages/Access/type.ts
@@ -1,0 +1,9 @@
+export enum ActiveKey {
+  ACCOUNTS = 'accounts',
+  ROLES = 'roles',
+}
+
+export enum Type {
+  CREATE = 'create',
+  EDIT = 'edit',
+}
